# public/swagger_old.json
{
	"swagger": "2.0",
	"info": {
		"version": "1.0.0",
		"title": "Brokering Portfolio"
	},
	"host": "localhost:3000",
...
	"schemes": [
		"http"
	],
	"consumes": [
		"application/json"
	],
	"produces": [
		"application/json"
	],
	"paths": {
...
	"/users/assetList": {
		"get": {
			"description": "Returns List of Assest holding by customers",
			"operationId": "GetAllAssets",
			"produces": [
				"application/json"
			],
			
			"responses": {
				"200": {
					"description": "Assest responses"
				},
				"default": {
					"description": "unexpected error"
				}
			}
		}
	},
...
	}
}